refactor result
